requirements.txt
streamlit==1.28.0
langchain==0.0.346
chromadb==0.4.14
sentence-transformers==2.2.2
transformers==4.35.0
torch>=2.1.0
pypdf2==3.0.1
pydantic==2.5.0
pillow==10.0.0
numpy>=1.26.0
matplotlib==3.7.0
wordcloud==1.9.0